Fixed some bugs and trying to implement creating new tracker, will continue tomorrow
