+++ Tracker/AddingHabits/ScheduleViewController.swift
@@ -4,12 +4,14 @@ import UIKit
 final class ScheduleViewController: UIViewController {
     // MARK: - Properties:
     weak var delegate: ScheduleViewControllerProtocol?
+    
     // MARK: - Private properties:
     private lazy var topTitle: UILabel = {
-        let label = UILabel()
-        label.text = "Расписание"
-        label.font = .systemFont(ofSize: 16, weight: .medium)
-        label.textColor = .YPBlack
+        let label                                       = UILabel()
+        label.text                                      = "Расписание"
+        label.font                                      = .systemFont(ofSize: 16, weight: .medium)
+        label.textColor                                 = .YPBlack
+        label.translatesAutoresizingMaskIntoConstraints = false
         
         return label
     }()
@@ -21,26 +23,43 @@ final class ScheduleViewController: UIViewController {
                                              "Суббота":     5,
                                              "Воскресенье": 6]
     
-    private let weekDays: [String] = [
-        "Понедельник", "Вторник", "Среда", "Четверг",
-        "Пятница", "Суббота", "Воскресенье"
-    ]
+    private let weekDays: [String] = ["Понедельник",
+                                      "Вторник",
+                                      "Среда",
+                                      "Четверг",
+                                      "Пятница",
+                                      "Суббота",
+                                      "Воскресенье"]
+    
     private var selectedDays: [String] = [] {
         didSet {
             selectedDays.sort { (dictToSort[$0] ?? 7) < (dictToSort[$1] ?? 7) }
         }
     }
-  
-    private let tableView = UITableView()
+    
+    private lazy var tableView: UITableView = {
+       let table                                        = UITableView()
+        table.delegate                                  = self
+        table.dataSource                                = self
+        table.backgroundColor                           = .YPBackground
+        table.layer.masksToBounds                       = true
+        table.layer.cornerRadius                        = 16
+        table.isScrollEnabled                           = false
+        table.translatesAutoresizingMaskIntoConstraints = false
+        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "Cell")
+        
+        return table
+    }()
     
     private lazy var doneButton: UIButton = {
-        let button = UIButton(type: .system)
+        let button                                       = UIButton(type: .system)
+        button.tintColor                                 = .YPWhite
+        button.frame.size                                = CGSize(width: 335, height: 60)
+        button.layer.masksToBounds                       = true
+        button.layer.cornerRadius                        = 16
+        button.backgroundColor                           = .YPBlack
+        button.translatesAutoresizingMaskIntoConstraints = false
         button.setTitle("Готово", for: .normal)
-        button.tintColor           = .YPWhite
-        button.frame.size          = CGSize(width: 335, height: 60)
-        button.layer.masksToBounds = true
-        button.layer.cornerRadius  = 16
-        button.backgroundColor = .YPBlack
         button.addTarget(self, action: #selector(doneButtonTapped), for: .touchUpInside)
         
         return button
@@ -49,6 +68,7 @@ final class ScheduleViewController: UIViewController {
     // MARK: - LifeCycle:
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
+        
         selectedDays = delegate?.selectedDays ?? []
         print(selectedDays)
     }
@@ -62,16 +82,10 @@ final class ScheduleViewController: UIViewController {
     
     // MARK: - Private methods:
     private func configureScreenItems() {
-        topTitle.translatesAutoresizingMaskIntoConstraints   = false
-        tableView.translatesAutoresizingMaskIntoConstraints  = false
-        doneButton.translatesAutoresizingMaskIntoConstraints = false
-        
         view.addSubview(topTitle)
         view.addSubview(tableView)
         view.addSubview(doneButton)
-        
-        tableViewConfig()
-        
+      
         NSLayoutConstraint.activate([
             topTitle.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 39),
             topTitle.centerXAnchor.constraint(equalTo: view.safeAreaLayoutGuide.centerXAnchor),
@@ -88,16 +102,6 @@ final class ScheduleViewController: UIViewController {
             doneButton.heightAnchor.constraint(equalToConstant: 60),
             doneButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -16),
         ])
-    }
-    
-    private func tableViewConfig() {
-        tableView.delegate   = self
-        tableView.dataSource = self
-        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "Cell")
-        tableView.backgroundColor = .YPBackground
-        tableView.layer.masksToBounds = true
-        tableView.layer.cornerRadius = 16
-        tableView.isScrollEnabled = false
     }
     
     @objc private func switchToggled(_ sender: UISwitch) {
@@ -124,7 +128,6 @@ extension ScheduleViewController: UITableViewDelegate {
     }
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let cell = tableView.cellForRow(at: indexPath)
         tableView.deselectRow(at: indexPath, animated: true)
         
     }
@@ -137,20 +140,21 @@ extension ScheduleViewController: UITableViewDataSource {
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "Cell")!
+        let cell                    = tableView.dequeueReusableCell(withIdentifier: "Cell")!
         
-        let switchView = UISwitch(frame: .zero)
-        switchView.tag = indexPath.row
-        switchView.onTintColor = .YPBlue
+        let switchView              = UISwitch(frame: .zero)
+        switchView.tag              = indexPath.row
+        switchView.onTintColor      = .YPBlue
         switchView.addTarget(self, action: #selector(switchToggled(_:)), for: .valueChanged)
-        cell.accessoryView = switchView
-        cell.textLabel?.text = weekDays[indexPath.row]
-        cell.backgroundColor = .clear
-        cell.separatorInset = UIEdgeInsets(top: 0, left: 16, bottom: 0, right: 16)
+        cell.accessoryView          = switchView
+        cell.textLabel?.text        = weekDays[indexPath.row]
+        cell.backgroundColor        = .clear
+        cell.separatorInset         = UIEdgeInsets(top: 0, left: 16, bottom: 0, right: 16)
         if indexPath.row == 6 {
-            cell.separatorInset = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 10000)
+            cell.separatorInset     = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 10000)
         }
-        let item = weekDays[indexPath.row]
+        let item                    = weekDays[indexPath.row]
+        
         for day in selectedDays {
             if day == item {
                 switchView.isOn = true
