+++ Tracker/Models/Tracker.swift
@@ -4,9 +4,10 @@ import UIKit
 struct Tracker {
     let id: UUID
     let name: String
-    let color: UIColor
-    let emoji: String
     let schedule: [String]
+//    let color: UIColor
+//    let emoji: String
+    
 }
