+++ Tracker/TabBar/TabBarViewController.swift
@@ -3,11 +3,11 @@ import UIKit
 final class TabBarViewController: UITabBarController {
     // MARK: - Private properties:
     let datePicker: UIDatePicker = {
-        let picker = UIDatePicker()
-        picker.preferredDatePickerStyle = .compact
-        picker.datePickerMode = .date
-        picker.tintColor = .YPBlue
-        picker.translatesAutoresizingMaskIntoConstraints = false
+        let picker                                                   = UIDatePicker()
+        picker.preferredDatePickerStyle                              = .compact
+        picker.datePickerMode                                        = .date
+        picker.tintColor                                             = .YPBlue
+        picker.translatesAutoresizingMaskIntoConstraints             = false
         picker.heightAnchor.constraint(equalToConstant: 34).isActive = true
         picker.widthAnchor.constraint(equalToConstant: 110).isActive = true
         picker.addTarget(self, action: #selector(datePickerValueChanged(_:)), for: .valueChanged)
@@ -30,14 +30,13 @@ final class TabBarViewController: UITabBarController {
     }
     
     private func createTrackersNav(with title: String, and image: UIImage?, from vc: UIViewController) -> UINavigationController {
-        let trackersNav = UINavigationController(rootViewController: vc)
-        trackersNav.tabBarItem.title = title
-        trackersNav.tabBarItem.image = image
+        let trackersNav                     = UINavigationController(rootViewController: vc)
+        trackersNav.tabBarItem.title        = title
+        trackersNav.tabBarItem.image        = image
         trackersNav.navigationBar.tintColor = .YPBlack
-        let rightButton = UIBarButtonItem(customView: datePicker)
+        let rightButton                     = UIBarButtonItem(customView: datePicker)
         trackersNav.viewControllers.first?.navigationItem.rightBarButtonItem = rightButton
-        trackersNav.viewControllers.first?.navigationItem.leftBarButtonItem = UIBarButtonItem(image: UIImage(named: "Plus"), style: .plain, target: self, action: #selector(createHabitOrEvent))
-        
+        trackersNav.viewControllers.first?.navigationItem.leftBarButtonItem  = UIBarButtonItem(image: UIImage(named: "Plus"), style: .plain, target: self, action: #selector(createHabitOrEvent))
         
         vc.navigationItem.title = title
         vc.navigationController?.navigationBar.prefersLargeTitles = true
@@ -46,24 +45,24 @@ final class TabBarViewController: UITabBarController {
     }
     
     private func createStatsNav(with title: String, and image: UIImage?, from vc: UIViewController) -> UINavigationController {
-        let statsNav = UINavigationController(rootViewController: vc)
+        let statsNav              = UINavigationController(rootViewController: vc)
         statsNav.tabBarItem.title = title
         statsNav.tabBarItem.image = image
-        vc.navigationItem.title = title
+        vc.navigationItem.title   = title
         vc.navigationController?.navigationBar.prefersLargeTitles = true
         
         return statsNav
     }
     
     @objc private func createHabitOrEvent() {
-        self.present(CreatingViewController(), animated: true)
+        self.present(HabitOrEventViewController(), animated: true)
     }
     
     @objc private func datePickerValueChanged(_ sender: UIDatePicker) {
-        let selectedDate = sender.date
-        let dateFormatter = DateFormatter()
+        let selectedDate         = sender.date
+        let dateFormatter        = DateFormatter()
         dateFormatter.dateFormat = "dd.MM.yyy"
-        let formattedDate = dateFormatter.string(from: selectedDate)
+        let formattedDate        = dateFormatter.string(from: selectedDate)
         print("Current date: \(formattedDate)")
     }
